requirements.txt
mysql-connector-python
numpy
datetime
matplotlib